requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
plivo==4.47.0
openai==1.12.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
aiohttp==3.9.3
numpy==1.26.3
